refactor : 숫자 비교 함수들과 야구게임 진행/재시작 함수들 묶어 분리
